refactor: move handler errors to own module
