+++ src/services/signer.services.ts
@@ -1,8 +1,12 @@
 import {isNullish} from '@dfinity/utils';
 import {consentMessage} from '../api/canister.api';
-import {SignerErrorCode} from '../constants/signer.constants';
 import type {icrc21_consent_info} from '../declarations/icrc-21';
-import {notifyError} from '../handlers/signer.handlers';
+import {
+  notifyErrorActionAborted,
+  notifyErrorRequestNotSupported,
+  notifyMissingPromptError,
+  notifyNetworkError
+} from '../handlers/signer-errors.handlers';
 import type {IcrcCallCanisterRequestParams} from '../types/icrc-requests';
 import type {Notify} from '../types/signer-handlers';
 import type {SignerOptions} from '../types/signer-options';
@@ -100,58 +104,4 @@ const promptConsentMessage = async ({
   });
 
   return await promise;
-};
-
-export const notifyErrorRequestNotSupported = ({
-  message,
-  ...notify
-}: Notify & {message?: string}): void => {
-  notifyError({
-    ...notify,
-    error: {
-      code: SignerErrorCode.REQUEST_NOT_SUPPORTED,
-      message: message ?? 'The request sent by the relying party is not supported by the signer.'
-    }
-  });
-};
-
-const notifyErrorActionAborted = (notify: Notify): void => {
-  notifyError({
-    ...notify,
-    error: {
-      code: SignerErrorCode.ACTION_ABORTED,
-      message: 'The signer has canceled the action requested by the relying party.'
-    }
-  });
-};
-
-const notifyNetworkError = ({message, ...notify}: Notify & {message: string}): void => {
-  notifyError({
-    ...notify,
-    error: {
-      code: SignerErrorCode.NETWORK_ERROR,
-      message
-    }
-  });
-};
-
-export const notifyErrorPermissionNotGranted = (notify: Notify): void => {
-  notifyError({
-    ...notify,
-    error: {
-      code: SignerErrorCode.PERMISSION_NOT_GRANTED,
-      message:
-        'The signer has not granted the necessary permissions to process the request from the relying party.'
-    }
-  });
-};
-
-export const notifyMissingPromptError = (notify: Notify): void => {
-  notifyError({
-    ...notify,
-    error: {
-      code: SignerErrorCode.PERMISSIONS_PROMPT_NOT_REGISTERED,
-      message: 'The signer has not registered a prompt to respond to permission requests.'
-    }
-  });
 };
